Fix Deprecation Warning to_float with tf.cast

+++ src/sample.py
@@ -62,7 +62,7 @@ def sample_sequence(*, hparams, length, start_token=None, batch_size=None, conte
 
         def body(past, prev, output):
             next_outputs = step(hparams, prev[:, tf.newaxis], past=past)
-            logits = next_outputs['logits'][:, -1, :]  / tf.to_float(temperature)
+            logits = next_outputs['logits'][:, -1, :]  / tf.cast(temperature, tf.float32)
             if top_p > 0.0:
                 logits = top_p_logits(logits, p=top_p)
             else:
